Optimize the music score for space.
Reduce the size of the score by a factor of almost 4 by noting that:

- We only use one channel in this game
- We only have 22 distinct notes
- We only have 7 distinct durations between notes

So we compress a 4-byte note-on and duration pair to one byte, with the
note in the lower 5 bits, and duration in the upper 3 bits. We now need
to add another array to map 3-bit duration codes to an actual interval,
but we also remove unused frequencies from the frequency table.

Note-off is just a note-on command with 0 for the note number.