Make the RGB LED color reflect the statistics.
The color of the LED is determined by the minimum value of the four
statistics. If the minimum value is 100%, the LED is green, and gradually
changes to red when the minimum values is 0%.

Use a custom setRGBled() function that uses all three output compare
registers of timer 1 and leaves timer 0 alone.